webapp/requirements.txt
fastapi>=0.112,<0.113
uvicorn[standard]>=0.30,<0.31
faiss-cpu>=1.8.0,<2.0
sentence-transformers>=3.0,<3.2
numpy>=1.26,<3.0
requests>=2.31,<3.0

